Renaming and moving files about seeding
